+++ packages/base-monaco-editor/src/index.tsx
@@ -11,6 +11,7 @@ import {
 } from './helper';
 
 export * from './monaco';
+export * from './controller';
 
 const SingleMonacoEditor = (props: ISingleMonacoEditorProps) => {
   const { onChange, enableOutline, width, height, language, supportFullScreen } = props;