+++ packages/base-monaco-editor/src/monaco.ts
@@ -1,18 +1,7 @@
 import loader, { Monaco } from '@monaco-editor/loader';
+//@ts-ignore
 import isEqual from 'lodash/isEqual';
-
-export interface Configure {
-  singleton?: boolean;
-}
-
-const CONFIGURE_KEY = '__base_monaco_editor_config__';
-const fakeWindow: any = window;
-
-if (!fakeWindow[CONFIGURE_KEY]) {
-  fakeWindow[CONFIGURE_KEY] = {};
-}
-
-export const configuration: Configure = fakeWindow[CONFIGURE_KEY];
+import { controller, EditorMeta } from './controller';
 
 export const getSingletonMonaco = (() => {
   let monaco: Monaco;
@@ -29,7 +18,9 @@ export const getSingletonMonaco = (() => {
               },
             },
       );
+      // eslint-disable-next-line require-atomic-updates
       monaco = await loader.init();
+      // eslint-disable-next-line require-atomic-updates
       prevOptions = options;
     }
     return monaco;
@@ -49,14 +40,14 @@ export const getCommonMonaco = (config: any): Promise<Monaco> => {
   return loader.init();
 };
 
-export function getMonaco(config: any) {
+export function getMonaco(config?: any) {
   const hasConfig = Object.keys(config || {}).length > 0;
   const monacoConfig = hasConfig ? config : undefined;
-  return configuration.singleton
+  return controller.getMeta().singleton
     ? getSingletonMonaco(monacoConfig)
     : getCommonMonaco(monacoConfig);
 }
 
-export function configure(conf: Configure) {
-  Object.assign(configuration, conf || {});
+export function configure(config: EditorMeta) {
+  controller.updateMeta(config);
 }